staging: android: logger: Add a TODO
A comment about a pre-existing bug data structure definition is added.
This bug was evidently introduced by Xiong Zhou in the patch
bd471258f2e09 ("staging: android: logger: use kuid_t instead of uid_t")
where the code was changed to hide a valid build warning instead of
solving the bug that was identified by it.